backend/pyproject.toml
[tool.poetry]
name = "doc-intelligence-system"
version = "1.0.0"
description = "Document Intelligence System Backend"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncmy = "^0.2.8"
alembic = "^1.13.0"
redis = "^5.0.0"
celery = "^5.3.0"
pdfplumber = "^0.10.0"
python-docx = "^1.1.0"
pandas = "^2.1.0"
openpyxl = "^3.1.0"
httpx = "^0.25.0"
tenacity = "^8.2.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
aiofiles = "^23.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
